chore: migrate to using the same approach as wp-core chore: add common config to the npm package to be used in theme & plugin
